+++ programs/drift/src/controller/orders/tests.rs
@@ -130,6 +130,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -249,6 +250,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -368,6 +370,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -487,6 +490,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -606,6 +610,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -691,6 +696,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -777,6 +783,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -863,6 +870,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -949,6 +957,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -1055,6 +1064,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -1164,6 +1174,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -1280,6 +1291,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -1397,6 +1409,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -1538,6 +1551,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -1654,6 +1668,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -1778,6 +1793,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut oracle_map,
+            false,
         )
         .unwrap();
 
@@ -1919,6 +1935,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut oracle_map,
+            false,
         )
         .unwrap();
 
@@ -2058,6 +2075,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut oracle_map,
+            false,
         )
         .unwrap();
 
@@ -2198,6 +2216,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut oracle_map,
+            false,
         )
         .unwrap();
 
@@ -2325,6 +2344,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -2451,6 +2471,7 @@ pub mod fulfill_order_with_maker_order {
             slot,
             &fee_structure,
             &mut get_oracle_map(),
+            false,
         )
         .unwrap();
 
@@ -5152,7 +5173,7 @@ pub mod fill_order {
             ..State::default()
         };
 
-        let base_asset_amount = fill_perp_order(
+        let (base_asset_amount, _) = fill_perp_order(
             1,
             &state,
             &user_account_loader,
@@ -5354,7 +5375,7 @@ pub mod fill_order {
             ..State::default()
         };
 
-        let base_asset_amount = fill_perp_order(
+        let (base_asset_amount, _) = fill_perp_order(
             1,
             &state,
             &user_account_loader,
@@ -5483,7 +5504,7 @@ pub mod fill_order {
             unix_timestamp: 11,
         };
 
-        let base_asset_amount = fill_perp_order(
+        let (base_asset_amount, _) = fill_perp_order(
             1,
             &state,
             &user_account_loader,