+++ programs/drift/src/lib.rs
@@ -412,6 +412,13 @@ pub mod drift {
         )
     }
 
+    pub fn liquidate_perp_with_fill<'c: 'info, 'info>(
+        ctx: Context<'_, '_, 'c, 'info, LiquidatePerp<'info>>,
+        market_index: u16,
+    ) -> Result<()> {
+        handle_liquidate_perp_with_fill(ctx, market_index)
+    }
+
     pub fn liquidate_spot<'c: 'info, 'info>(
         ctx: Context<'_, '_, 'c, 'info, LiquidateSpot<'info>>,
         asset_market_index: u16,
@@ -823,6 +830,13 @@ pub mod drift {
         amount: u64,
     ) -> Result<()> {
         handle_deposit_into_perp_market_fee_pool(ctx, amount)
+    }
+
+    pub fn deposit_into_spot_market_vault<'c: 'info, 'info>(
+        ctx: Context<'_, '_, 'c, 'info, DepositIntoSpotMarketVault<'info>>,
+        amount: u64,
+    ) -> Result<()> {
+        handle_deposit_into_spot_market_vault(ctx, amount)
     }
 
     pub fn deposit_into_spot_market_revenue_pool<'c: 'info, 'info>(